backend/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Configuration
python-dotenv>=1.0.0

# HTTP client
httpx>=0.25.2
requests>=2.31.0

# Database
sqlalchemy>=2.0.23
alembic>=1.12.1
psycopg2-binary>=2.9.9

# External APIs
spotipy>=2.23.0
openai>=1.3.8

# Data processing (Python 3.13 compatible versions)
pandas>=2.2.0
numpy>=1.26.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1

# Settings validation
pydantic-settings>=2.0.0
